requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.34
python-dotenv==1.0.0
Werkzeug==2.3.7